refactor(client): user FollowingSearch and FollowersSearch refactor & move LoadingScreen folder
